steps: Move library code that will be shared to a new shared location
Moves the library code (helper functions etc.) to a new python module
called `common` in the behave root directory. This code can then be
shared between various test suites.

Adds the path to behave root directory to python path in environment.py,
so that the `common` module can be imported from the steps code.

The new directory structure is as follows:
common
├── <common step files>
├── lib
│   └── <shared lib code (previously in common)>
features
└── steps
    ├── <dnf-specific steps>
    └── lib
        └── <dnf-specific lib code (previously in common)>
