+++ dnf-behave-tests/common/lib/checksum.py
