+++ dnf-behave-tests/common/lib/diff.py
