+++ dnf-behave-tests/features/steps/file.py
@@ -8,20 +8,20 @@ import glob
 import re
 import os
 
-from common.behave_ext import check_context_table
-from common.checksum import sha256_checksum
-from common.cmd import run
-from common.file import create_file_with_contents
-from common.file import copy_file
-from common.file import copy_tree
-from common.file import delete_directory
-from common.file import delete_file
-from common.file import ensure_directory_exists
-from common.file import ensure_file_exists
-from common.file import file_timestamp
-from common.file import prepend_installroot
-from common.file import read_file_contents
-from common.string import print_lines_diff
+from common.lib.behave_ext import check_context_table
+from common.lib.checksum import sha256_checksum
+from common.lib.cmd import run
+from common.lib.file import create_file_with_contents
+from common.lib.file import copy_file
+from common.lib.file import copy_tree
+from common.lib.file import delete_directory
+from common.lib.file import delete_file
+from common.lib.file import ensure_directory_exists
+from common.lib.file import ensure_file_exists
+from common.lib.file import file_timestamp
+from common.lib.file import prepend_installroot
+from common.lib.file import read_file_contents
+from common.lib.diff import print_lines_diff
 
 
 @behave.given('I create directory "{dirpath}"')
